+++ plone/app/contentrules/handlers.py
@@ -1,13 +1,12 @@
 import threading
 
 from zope.component import queryUtility
-from zope.container.interfaces import IObjectAddedEvent, IObjectRemovedEvent,\
+from zope.container.interfaces import IObjectAddedEvent, IObjectRemovedEvent, \
     IContainerModifiedEvent
 from zope.interface import Interface
 from zope.component.hooks import getSite
 
 from plone.app.discussion.interfaces import IComment
-from plone.contentrules.rule.interfaces import IRule
 from plone.contentrules.engine.interfaces import IRuleExecutor
 from plone.contentrules.engine.interfaces import IRuleStorage
 from plone.contentrules.engine.interfaces import StopRule
@@ -16,8 +15,6 @@ from Acquisition import aq_inner, aq_parent
 from plone.uuid.interfaces import IUUID
 from Products.CMFCore.interfaces import ISiteRoot, IContentish
 from Products.CMFCore.utils import getToolByName
-from Products.ZCTextIndex.interfaces import IZCLexicon
-from AccessControl.userfolder import UserFolder
 
 try:
     from Products.Archetypes.interfaces import IBaseObject
@@ -26,6 +23,8 @@ try:
 except ImportError:
     class IBaseObject(Interface):
         pass
+
+
     class IObjectInitializedEvent(Interface):
         pass
     HAS_ARCHETYPES = False
@@ -140,9 +139,7 @@ def execute(context, event):
     # execute rules again
     rule_filter.in_progress = False
 
-
 # Event handlers
-
 
 def is_portal_factory(context):
     """Find out if the given object is in portal_factory