+++ plone/app/contentrules/namechooser.py
@@ -2,7 +2,6 @@ from zope.container.contained import NameChooser
 from zope.container.interfaces import INameChooser
 from zope.interface import implements
 
-
 ATTEMPTS = 100