+++ plone/app/contentrules/tests/base.py
@@ -1,36 +1,26 @@
-"""Base class for integration tests, based on ZopeTestCase and PloneTestCase.
-
-Note that importing this module has various side-effects: it registers a set of
-products with Zope, and it sets up a sandbox Plone site with the appropriate
-products installed.
+"""Base class for integration tests, based on plone.app.testing
 """
 
-# Import PloneTestCase - this registers more products with Zope as a side effect
-from Products.PloneTestCase.PloneTestCase import PloneTestCase
-from Products.PloneTestCase.PloneTestCase import FunctionalTestCase
-from Products.PloneTestCase.PloneTestCase import setupPloneSite
+from plone.app.testing.bbb import PloneTestCase
 from zope.component import getMultiAdapter
-
-# Set up a Plone site - note that the portlets branch of CMFPlone applies
-# a portlets profile.
-setupPloneSite()
 
 
 class ContentRulesTestCase(PloneTestCase):
-    """Base class for integration tests for plone.app.contentrules. This may
-    provide specific set-up and tear-down operations, or provide convenience
-    methods.
+    """Base class for integration tests for plone.app.contentrules.
+    This may provide specific set-up and tear-down operations, or provide
+    convenience methods.
     """
 
     def addAuthToRequest(self):
-        portal = self.portal
-        request = portal.REQUEST
+        portal = self.layer['portal']
+        request = self.layer['request']
         authenticator = getMultiAdapter((portal, request), name=u"authenticator")
         auth = authenticator.authenticator().split('value="')[1].rstrip('"/>')
         request.form['_authenticator'] = auth
 
-class ContentRulesFunctionalTestCase(FunctionalTestCase):
-    """Base class for functional integration tests for plone.app.portlets.
+
+class ContentRulesFunctionalTestCase(PloneTestCase):
+    """Base class for functional integration tests for plone.app.contentrules.
     This may provide specific set-up and tear-down operations, or provide
     convenience methods.
     """