+++ plone/app/contentrules/tests/test_action_modify.py
@@ -38,10 +38,3 @@ class TestModifyAction(TestCase):
             __name__ = None
         handlers.modified(ObjectRemovedEvent(Content()))
         self.assertFalse(self.called)
-
-
-def test_suite():
-    from unittest import TestSuite, makeSuite
-    suite = TestSuite()
-    suite.addTest(makeSuite(TestModifyAction))
-    return suite