+++ plone/app/contentrules/tests/test_cascading_rule.py
@@ -1,28 +1,10 @@
 # -*- coding: utf-8 -*-
-
-import time
-
-from plone.contentrules.engine.interfaces import IRuleAssignmentManager
 from plone.contentrules.engine.interfaces import IRuleStorage
 from zope.component import getUtility
-from zope.component import getMultiAdapter
-from zope.lifecycleevent.interfaces import IObjectModifiedEvent
-
-from Products.GenericSetup.interfaces import IBody
-from Products.GenericSetup.context import TarballExportContext
-from Products.PloneTestCase.layer import PloneSite
 
 from plone.app.contentrules.tests.base import ContentRulesTestCase
 from plone.app.contentrules.tests.test_configuration import TestContentrulesGSLayer
 from plone.app.contentrules.api import edit_rule_assignment
-
-# BBB Zope 2.12
-try:
-    from Zope2.App import zcml
-    from OFS import metaconfigure
-except ImportError:
-    from Products.Five import zcml
-    from Products.Five import fiveconfigure as metaconfigure
 
 
 class TestCascadingRule(ContentRulesTestCase):
@@ -55,12 +37,5 @@ class TestCascadingRule(ContentRulesTestCase):
         self.assertTrue('my-event' in self.portal.events)
 
         wtool = self.portal.portal_workflow
-        self.assertTrue(wtool.getInfoFor(self.portal.events['my-event'], 'review_state'),
+        self.assertEqual(wtool.getInfoFor(self.portal.events['my-event'], 'review_state'),
                         'published')
-
-
-def test_suite():
-    from unittest import TestSuite, makeSuite
-    suite = TestSuite()
-    suite.addTest(makeSuite(TestCascadingRule))
-    return suite