+++ plone/app/contentrules/tests/test_rule_management_views.py
@@ -138,7 +138,6 @@ class TestRuleElementManagementViews(ContentRulesTestCase):
         registered_rules = controlpanel.registeredRules()
         self.assertEqual(0, len(registered_rules))
 
-
     def testChangeGloballyEnable(self):
         storage = getUtility(IRuleStorage)
         portal = self.portal
@@ -155,10 +154,3 @@ class TestRuleElementManagementViews(ContentRulesTestCase):
 
         # without ajax
         self.portal.REQUEST.form
-
-def test_suite():
-    from unittest import TestSuite, makeSuite
-    suite = TestSuite()
-    suite.addTest(makeSuite(TestRuleManagementViews))
-    suite.addTest(makeSuite(TestRuleElementManagementViews))
-    return suite