+++ plone/app/contentrules/tests/test_setup.py
@@ -18,10 +18,3 @@ class TestProductInstall(ContentRulesTestCase):
         self.assertTrue(IRuleEventType.providedBy(IObjectAddedEvent))
         self.assertTrue(IRuleEventType.providedBy(IObjectModifiedEvent))
         self.assertTrue(IRuleEventType.providedBy(IObjectRemovedEvent))
-
-
-def test_suite():
-    from unittest import TestSuite, makeSuite
-    suite = TestSuite()
-    suite.addTest(makeSuite(TestProductInstall))
-    return suite